pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lbt_nrm_sim"
version = "0.1.0"
description = "LBT NRM aperture masking fringe simulator"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "Pillow",
]

[project.optional-dependencies]
dev = ["pytest", "ipython"]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"lbt_nrm_sim" = ["filters/*.dat"]
